Add Checkstyle
Add checkstyle to assist in coding style

to ensure a easier time coding and maintaining to conding
style standards, checkstyle helps catches potential
style errors that may slip our checks

automated process of checkstyle helps with speed
of coding
