+++ build.gradle
@@ -5,6 +5,10 @@ plugins {
     id 'com.github.johnrengelman.shadow' version '7.1.2'
 }
 
+checkstyle {
+    toolVersion = '10.2'
+}
+
 repositories {
     mavenCentral()
 }
