--- src/main/java/Exception/InvalidInputException.java
@@ -1,10 +1,0 @@
-package Exception;
-
-/**
- * Exception for handling Invalid command inputs
- */
-public class InvalidInputException extends Exception {
-  public InvalidInputException(String msg) {
-    super(msg);
-  }
-}
