--- src/main/java/Exception/StorageSyntaxException.java
@@ -1,10 +1,0 @@
-package Exception;
-
-/**
- * Exception for handling invalid storage format
- */
-public class StorageSyntaxException extends InvalidInputException{
-    public StorageSyntaxException() {
-        super("Apologies Commander, there appear to be an issue with the storage file");
-    }
-}
