--- src/main/java/Exception/SyntaxException.java
@@ -1,10 +1,0 @@
-package Exception;
-
-/**
- * Exception for handling Syntax Error in Commands
- */
-public class SyntaxException extends InvalidInputException {
-  public SyntaxException(String type, String style) {
-    super("Apologies Commander, invalid " + type + " format. Please use (" + style + ").");
-  }
-}
