+++ src/main/java/exception/InvalidInputException.java
@@ -1,0 +1,10 @@
+package exception;
+
+/**
+ * Exception for handling Invalid command inputs
+ */
+public class InvalidInputException extends Exception {
+  public InvalidInputException(String msg) {
+    super(msg);
+  }
+}
